% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{centenarian_df}
\alias{centenarian_df}
\title{Japanese Centenarians Data}
\format{
A data frame with 21 observations and 19 variables:
\describe{
  \item{X}{Age or identifier (numeric)}
  \item{X1963}{Number of deaths in 1963 (integer)}
  \item{X1964}{Number of deaths in 1964 (integer)}
  \item{X1965}{Number of deaths in 1965 (integer)}
  \item{X1966}{Number of deaths in 1966 (integer)}
  \item{X1967}{Number of deaths in 1967 (integer)}
  \item{X1968}{Number of deaths in 1968 (integer)}
  \item{X1969}{Number of deaths in 1969 (integer)}
  \item{X1970}{Number of deaths in 1970 (integer)}
  \item{X1971}{Number of deaths in 1971 (integer)}
  \item{X1972}{Number of deaths in 1972 (integer)}
  \item{X1973}{Number of deaths in 1973 (integer)}
  \item{X1974}{Number of deaths in 1974 (integer)}
  \item{X1975}{Number of deaths in 1975 (integer)}
  \item{X1976}{Number of deaths in 1976 (integer)}
  \item{X1977}{Number of deaths in 1977 (integer)}
  \item{X1978}{Number of deaths in 1978 (integer)}
  \item{X1979}{Number of deaths in 1979 (integer)}
  \item{X1980}{Number of deaths in 1980 (integer)}
}
}
\source{
Data taken from the depend.truncation package version 3.0
}
\usage{
data(centenarian_df)
}
\description{
This dataset, centenarian_df, is a data frame containing the number of deaths at each year
from 1963 to 1980 for Japanese male centenarians. The data is based on Table 2 of Emura and Murotani (2015),
and also references the original work of Sibuya & Hanayama (2004).
}
\details{
The dataset name has been kept as 'centenarian_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
JapanAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
