% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/japanapis-package.R
\docType{package}
\name{JapanAPIs}
\alias{JapanAPIs}
\alias{JapanAPIs-package}
\title{JapanAPIs: Access Japanese Data via Public APIs and Curated Datasets}
\description{
{
This package provides functions to access data from public RESTful APIs including 'Nager.Date', 'World Bank API', and 'REST Countries API', retrieving real-time or historical data related to Japan, such as holidays, economic indicators, international demographic and geopolitical indicators. Additionally, the package includes one of the largest  curated collections of open datasets focused on Japan, covering topics such as natural disasters, economic production, vehicle industry, air quality, demographics, and administrative divisions.
}
}
\details{
JapanAPIs: Access Japanese Data via Public APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access Japanese Data via Public APIs and Curated Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/japanapis}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
