% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{jpn_gdp_cons_df}
\alias{jpn_gdp_cons_df}
\title{Household Consumption and GDP in Japan (1978–2007)}
\format{
A data frame with 30 observations and 3 variables:
\describe{
  \item{Year}{Calendar year (integer)}
  \item{C}{Household Consumption (integer)}
  \item{Y}{Gross Domestic Product (GDP) (integer)}
}
}
\source{
Data taken from the SupMZ package version 0.2.0
}
\usage{
data(jpn_gdp_cons_df)
}
\description{
This dataset, jpn_gdp_cons_df, is a data frame containing information on Household Consumption (C)
and Gross Domestic Product (GDP, denoted as Y) in Japan from 1978 to 2007.
The data is useful for analyzing structural breaks under heteroskedasticity.
}
\details{
The dataset name has been kept as 'jpn_gdp_cons_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
JapanAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
