% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{jpn_us_cars_df}
\alias{jpn_us_cars_df}
\title{Attributes of Some US and Japanese Automobiles}
\format{
A data frame with 45 observations and 4 variables:
\describe{
  \item{Model}{Car model (factor)}
  \item{Country}{Country of origin (factor)}
  \item{Mileage}{Mileage in miles per gallon (integer)}
  \item{Price}{Price in US dollars (integer)}
}
}
\source{
Data taken from the fma package version 2.5
}
\usage{
data(jpn_us_cars_df)
}
\description{
This dataset, jpn_us_cars_df, is a data frame containing information on 45 automobiles
from the United States and Japan. It includes attributes such as model name, country
of origin, mileage, and price.
}
\details{
The dataset name has been kept as 'jpn_us_cars_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
JapanAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
