% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{jpn_usd_exchange_df}
\alias{jpn_usd_exchange_df}
\title{Yen-Dollar Exchange Rate (Weekly, 1975–1989)}
\format{
A data frame with 778 observations and 4 variables:
\describe{
  \item{date}{Time index of the observation (integer)}
  \item{s}{Spot exchange rate (numeric)}
  \item{f}{Forward exchange rate (numeric)}
  \item{s30}{30-day forward exchange rate (numeric)}
}
}
\source{
Data taken from the Ecdat package version 0.4-2
}
\usage{
data(jpn_usd_exchange_df)
}
\description{
This dataset, jpn_usd_exchange_df, is a data frame containing weekly observations of
the yen-dollar exchange rate from 1975 to 1989. It includes spot and forward rates,
along with 30-day forward rates.
}
\details{
The dataset name has been kept as 'jpn_usd_exchange_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
JapanAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
