% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{jpn_vehicle_prod_ts}
\alias{jpn_vehicle_prod_ts}
\title{Japanese Motor Vehicle Production (1947–1989)}
\format{
A univariate time series with 43 observations:
\describe{
  \item{Time}{Yearly observations from 1947 to 1989}
  \item{Values}{Motor vehicle production in Japan (in thousands)}
}
}
\source{
Data taken from the fma package version 2.5
}
\usage{
data(jpn_vehicle_prod_ts)
}
\description{
This dataset, jpn_vehicle_prod_ts, is a univariate time series containing the number of motor vehicles
produced in Japan from 1947 to 1989. The figures are recorded annually and expressed in thousands.
}
\details{
The dataset name has been kept as 'jpn_vehicle_prod_ts' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
JapanAPIs package and assists users in identifying its specific characteristics.
The suffix 'ts' indicates that the dataset is a time series object. The original content has not been modified
in any way.
}
