% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{jpnprefs_tbl_df}
\alias{jpnprefs_tbl_df}
\title{Prefectural Informations in Japan}
\format{
A tibble with 47 observations and 5 variables:
\describe{
  \item{jis_code}{JIS code of the prefecture (character)}
  \item{prefecture_kanji}{Prefectural name in kanji (character)}
  \item{prefecture}{Prefectural name in romaji (character)}
  \item{region}{Geographical region of the prefecture (character)}
  \item{major_island}{Major island the prefecture belongs to (character)}
}
}
\source{
Data taken from the zipangu package version 0.3.3
}
\usage{
data(jpnprefs_tbl_df)
}
\description{
This dataset, jpnprefs_tbl_df, is a tibble containing information about the 47 prefectures of Japan.
It includes details such as prefectural names in kanji and romaji, corresponding JIS codes,
the region each prefecture belongs to, and the major island associated with each.
}
\details{
The dataset name has been kept as 'jpnprefs_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
JapanAPIs package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble object. The original content has not been modified
in any way.
}
