% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{kobe_quake_1995_ts}
\alias{kobe_quake_1995_ts}
\title{1995 Kobe Earthquake Data}
\format{
A time series with 3,048 observations:
\describe{
  \item{observations}{Time series data related to the 1995 Kobe earthquake (numeric)}
}
}
\source{
Data taken from the waveslim package version 1.8.5
}
\usage{
data(kobe_quake_1995_ts)
}
\description{
This dataset, `kobe_quake_1995_ts`, is a time series containing data related to
the 1995 Kobe earthquake. The data are organized sequentially with a frequency of 1,
and consist of 3,048 observations.
}
\details{
The dataset name has been kept as 'kobe_quake_1995_ts' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
JapanAPIs package and assists users in identifying its specific characteristics.
The suffix 'ts' indicates that the dataset is a time series object. The original content has not been modified
in any way.
}
