% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{nikkei_stock_index_df}
\alias{nikkei_stock_index_df}
\title{Japanese NIKKEI Stock Index}
\format{
A data frame with 4,246 observations and 2 variables:
\describe{
  \item{index}{Date or index label (character)}
  \item{value}{Daily log return in percent (numeric)}
}
}
\source{
Data taken from the tsgarch package version 1.0.3
}
\usage{
data(nikkei_stock_index_df)
}
\description{
This dataset, nikkei_stock_index_df, is a data frame containing the daily log returns in percent
of the NIKKEI stock index for the period from 1984-01-04 to 2000-12-22.
}
\details{
The dataset name has been kept as 'nikkei_stock_index_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
JapanAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
