% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{tokaido_stations_tbl_df}
\alias{tokaido_stations_tbl_df}
\title{East Japan Railway's Tokaido Line Stations}
\format{
A tibble with 20 observations and 2 variables:
\describe{
  \item{st_code}{Station code (character)}
  \item{st_name}{Station name (character)}
}
}
\source{
Data taken from the ssrn package version 0.1.0
}
\usage{
data(tokaido_stations_tbl_df)
}
\description{
This dataset, tokaido_stations_tbl_df, is a tibble containing information about
stations on the Tokaido Line operated by East Japan Railway. It includes each station's
code and name.
}
\details{
The dataset name has been kept as 'tokaido_stations_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
JapanAPIs package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble object. The original content has not been modified
in any way.
}
