% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{tokyo_rainfall_vec}
\alias{tokyo_rainfall_vec}
\title{Tokyo Rainfall Data}
\format{
A numeric vector with 366 observations:
\describe{
  Daily rainfall measurements (numeric values ranging from 0 to 2)
}
}
\source{
Data taken from the gamlss.data package version 6.0-6
}
\usage{
data(tokyo_rainfall_vec)
}
\description{
This dataset, tokyo_rainfall_vec, is a numeric vector containing Tokyo rainfall data
from Kitagawa (1987), analysed also by Rue and Held (2005) and Fahrmeir and Tutz (2013).
It includes daily rainfall measurements over a period of 366 days.
}
\details{
The dataset name has been kept as 'tokyo_rainfall_vec' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
JapanAPIs package and assists users in identifying its specific characteristics.
The suffix 'vec' indicates that the dataset is a vector object. The original content has not been modified
in any way.
}
