% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{toyota_stock_prices_df}
\alias{toyota_stock_prices_df}
\title{Toyota Stock Prices}
\format{
A data frame with 255 observations and 2 variables:
\describe{
  \item{Month}{Month number from the beginning of the time series (integer)}
  \item{Value}{Average monthly stock price (numeric)}
}
}
\source{
Data taken from the MMAC package version 0.1.2
}
\usage{
data(toyota_stock_prices_df)
}
\description{
This dataset, toyota_stock_prices_df, is a data frame containing the average monthly stock prices
of Toyota from 1982 to 1998.
}
\details{
The dataset name has been kept as 'toyota_stock_prices_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
JapanAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
